requirements.txt
click==6.7
cycler==0.10.0
environ==1.0
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.2
numpy==1.19.0
pandas==1.0.5
psycopg2-binary==2.8.5
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
Werkzeug==0.14.1
